Break out persistence tests
